# swagger.json
{
	"components": {
...
		"schemas": {
...
			"IFileDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"type",
					"url",
					"size",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EPriority": {
				"enum": [
					"highest",
					"high",
					"medium",
					"low",
					"lowest"
				],
				"type": "string"
			},
			"IPriorityDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/EPriority"
					},
					"icon": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"icon",
					"color",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IPriorityListDto": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/IPriorityDto"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPriorityCreateRequest": {
				"properties": {
					"name": {
						"$ref": "#/components/schemas/EPriority"
					},
					"icon": {
						"type": "string"
					},
					"color": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"icon",
					"color"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPriorityUpdateRequest": {
				"properties": {
					"name": {
						"$ref": "#/components/schemas/EPriority"
					},
					"icon": {
						"type": "string"
					},
					"color": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IProfileDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"birthDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"gender": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"lastOnline": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"avatar": {
						"allOf": [
							{
								"$ref": "#/components/schemas/IFileDto"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"avatar"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProfileUpdateRequest": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"birthDate": {
						"type": "string",
						"format": "date-time"
					},
					"gender": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IProfileListDto": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/IProfileDto"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.0",
		"description": "REST API Template App",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Epifanov Andrei",
			"email": "epifanovmd@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/file/{id}": {
			"delete": {
				"operationId": "DeleteFile",
				"responses": {
					"200": {
						"description": "Количество удаленных файлов (обычно 1, если файл существует)",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"description": "Удалить файл.\nЭтот эндпоинт позволяет пользователю удалить файл по его ID. Доступ разрешен только пользователю, который загрузил файл, либо администратору.",
				"summary": "Удаление файла",
				"tags": [
					"Files"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID файла, который нужно удалить",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/priority/list": {
			"get": {
				"operationId": "GetPriorities",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPriorityListDto"
								}
							}
						}
					}
				},
				"tags": [
					"Priority"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/priority/{id}": {
			"get": {
				"operationId": "GetPriorityById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPriorityDto"
								}
							}
						}
					}
				},
				"tags": [
					"Priority"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdatePriority",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPriorityDto"
								}
							}
						}
					}
				},
				"tags": [
					"Priority"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPriorityUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletePriority",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Priority"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/priority/name/{name}": {
			"get": {
				"operationId": "GetPriorityByName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPriorityDto"
								}
							}
						}
					}
				},
				"tags": [
					"Priority"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/priority": {
			"post": {
				"operationId": "CreatePriority",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPriorityDto"
								}
							}
						}
					}
				},
				"tags": [
					"Priority"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPriorityCreateRequest"
							}
						}
					}
				}
			}
		},
		"/api/profile/my": {
			"get": {
				"operationId": "GetMyProfile",
				"responses": {
					"200": {
						"description": "Пользователь",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfileDto"
								}
							}
						}
					}
				},
				"description": "Получить профиль текущего пользователя.\nЭтот эндпоинт позволяет получить данные профиля пользователя, который выполнил запрос.\nИспользуется для получения информации о текущем пользователе, например, его имени, email, и других данных.",
				"summary": "Получение профиля текущего пользователя",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/profile/my/update": {
			"patch": {
				"operationId": "UpdateMyProfile",
				"responses": {
					"200": {
						"description": "Обновленный профиль пользователя",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfileDto"
								}
							}
						}
					}
				},
				"description": "Обновить профиль текущего пользователя.\nЭтот эндпоинт позволяет пользователю обновить свои данные, такие как имя, email и другие параметры профиля.",
				"summary": "Обновление профиля текущего пользователя",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Обновленные данные профиля",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProfileUpdateRequest",
								"description": "Обновленные данные профиля"
							}
						}
					}
				}
			}
		},
		"/api/profile/my/delete": {
			"delete": {
				"operationId": "DeleteMyProfile",
				"responses": {
					"200": {
						"description": "Сообщение об успешном удалении",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "Удалить профиль текущего пользователя.\nЭтот эндпоинт позволяет пользователю удалить свой профиль из системы.",
				"summary": "Удаление профиля текущего пользователя",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/profile/list": {
			"get": {
				"operationId": "GetProfiles",
				"responses": {
					"200": {
						"description": "Список всех профилей с информацией о них",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfileListDto"
								}
							}
						}
					}
				},
				"description": "Получить все профили.\nЭтот эндпоинт позволяет администраторам получить список всех пользователей системы.\nОн поддерживает пагинацию через параметры `offset` и `limit`.",
				"summary": "Получение всех профилей",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Смещение для пагинации",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Лимит количества возвращаемых профилей",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/profile/{userId}": {
			"get": {
				"operationId": "GetProfileById",
				"responses": {
					"200": {
						"description": "Пользователь по ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfileDto"
								}
							}
						}
					}
				},
				"description": "Получить профиль по ID.\nЭтот эндпоинт позволяет получить профиль другого пользователя по его ID. Доступен только для администраторов.",
				"summary": "Получение профиля по ID",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID пользователя, профиль которого нужно получить",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/profile/update/{userId}": {
			"patch": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Обновленный профиль пользователя",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfileDto"
								}
							}
						}
					}
				},
				"description": "Обновить профиль другого пользователя.\nЭтот эндпоинт позволяет администраторам обновлять профиль других пользователей.",
				"summary": "Обновление профиля другого пользователя",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"description": "ID пользователя, профиль которого необходимо обновить",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Данные для обновления профиля",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProfileUpdateRequest",
								"description": "Данные для обновления профиля"
							}
						}
					}
				}
			}
		},
		"/api/profile/avatar/upload": {
			"post": {
				"operationId": "AddAvatar",
				"responses": {
					"200": {
						"description": "Обновленный профиль с новым аватаром",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfileDto"
								}
							}
						}
					}
				},
				"description": "Загрузить аватар для текущего пользователя.\nЭтот эндпоинт позволяет пользователю загрузить аватар для своего профиля.",
				"summary": "Загрузка аватара",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary",
										"description": "Файл изображения аватара"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
		"/api/profile/avatar": {
			"delete": {
				"operationId": "RemoveAvatar",
				"responses": {
					"200": {
						"description": "Обновленный профиль без аватара",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfileDto"
								}
							}
						}
					}
				},
				"description": "Удалить аватар пользователя.\nЭтот эндпоинт позволяет пользователю удалить свой аватар.",
				"summary": "Удаление аватара",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/profile/delete/{userId}": {
			"delete": {
				"operationId": "DeleteProfile",
				"responses": {
					"200": {
						"description": "Сообщение об успешном удалении",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "Удалить профиль другого пользователя.\nЭтот эндпоинт позволяет администраторам удалить профиль другого пользователя из системы.",
				"summary": "Удаление профиля другого пользователя",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"description": "ID пользователя, профиль которого необходимо удалить",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8181",
			"description": "Local dev"
		}
	]
}